pyproject.toml
[project]
name = "predmaint-ml-platform"
version = "0.1.0"
description = "End-to-End Predictive Maintenance with Automated Retraining, Drift Detection & Production Monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.66",
    "evidently>=0.7.20",
    "fastapi>=0.115.0",
    "joblib>=1.5.3",
    "mlflow>=3.10.1",
    "pandas>=2.2.0,<3",
    "plotly>=5.24.1",
    "prefect>=3.6.21",
    "pyarrow>=14.0.0,<24",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.34.0",
    "xgboost>=2.1.0",
]

[dependency-groups]
dev = [
    "boto3-stubs[s3]>=1.42.66",
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
    "scikit-learn-stubs>=0.0.3",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20260107",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning:evidently",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # assert statements are fine in tests (excluded) and not in src

[tool.mypy]
python_version = "3.12"
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "joblib",
    "evidently",
    "evidently.*",
    "plotly",
    "plotly.*",
    "mlflow.xgboost",
    "prefect",
    "prefect.*",
]
ignore_missing_imports = true
